src/requirements.txt
asgiref==3.5.2
Django==4.0.6
django-debug-toolbar==3.5.0
Pillow==9.2.0
sqlparse==0.4.2
django-jalali-date==1.0.1
jdatetime==4.1.0
